+++ jest.config.js
@@ -14,7 +14,7 @@ const jestConfig = {
   testEnvironment: 'node', // Use Node.js environment for testing
   moduleNameMapper: { '^(\\.{1,2}/.*)\\.js$': '$1' }, // Handle ESM imports by removing the .js extension
   testPathIgnorePatterns: ['/node_modules/', '/dist/', '/vitest/', '/frontend/', '/src/crypto/', '/src/mock/', 'jestHelpers.ts'], // Ignore specific paths for test files
-  coveragePathIgnorePatterns: ['/node_modules/', '/dist/', '/vitest/', '/frontend/', '/src/crypto/', '/src/mock/', 'jestHelpers.ts'], // Ignore specific paths for test and coverage
+  coveragePathIgnorePatterns: ['/node_modules/', '/dist/', '/vitest/', '/frontend/', '/src/crypto/', '/src/mock/', 'jestHelpers.ts', 'matterNode.ts'], // Ignore specific paths for test and coverage
   maxWorkers: '100%', // Use all available CPU cores for running tests
 };