+++ src/frontend.test.ts
@@ -138,6 +138,7 @@ describe('Matterbridge frontend', () => {
       dst: 'frontend',
       params: { plugin: 'test', serialNumber: '1234', uniqueId: 'uniqueId', number: 123, id: 'id', cluster: 'cluster', attribute: 'attribute', value: 'value' },
     } as any);
+    await (frontend as any).msgHandler({ id: 123456, type: 'frontend_logmessage', src: 'manager', dst: 'frontend', params: { level: 'info', time: 'time', name: 'jest', message: 'info' } } as any);
     for (const type of ['plugins_install', 'plugins_uninstall'] as const) {
       await (frontend as any).msgHandler({ id: 123456, type, src: 'manager', dst: 'all', response: { success: true, packageName: 'testPlugin' } } as any);
       await (frontend as any).msgHandler({ id: 123456, type, src: 'manager', dst: 'all', response: { success: false, packageName: 'testPlugin' } } as any);